- Removing the ignoreTimeout param from the command attribute and reworking some of the code to adjust for that. Tests are broken at this time and need some rework.
